Turned CompiledSystem into a proper system with inputs and outputs and extended the simulator to support a function for providing the inputs to the simulated system.
The outputs are now exposed separately from the signals in a CompiledSystem, so that interesting signals can be explicitly exposed by assigning appropriate output ports in the root block.

Also, the root block may now have inputs, and the simulator will honor a callable to provide inputs to the root block during simulation.